style: Rename package allocate to diskalloc
