+++ internal/pkg/service/buffer/sink/tablesink/storage/file_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/key"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/compression"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/allocate"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/diskalloc"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/disksync"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/staging"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/target"
@@ -94,7 +94,7 @@ func TestFile_Validation(t *testing.T) {
 		Dir:            "my-dir",
 		Compression:    compression.NewConfig(),
 		DiskSync:       disksync.NewConfig(),
-		DiskAllocation: allocate.NewConfig(),
+		DiskAllocation: diskalloc.NewConfig(),
 	}
 	stagingStorage := staging.File{
 		Compression:                 compression.NewConfig(),
