+++ internal/pkg/service/buffer/sink/tablesink/storage/level/local/config.go
@@ -2,7 +2,7 @@ package local
 
 import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/compression"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/allocate"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/diskalloc"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/disksync"
 )
 
@@ -13,7 +13,7 @@ type Config struct {
 	// DiskSync configures the synchronization of the in-memory copy of written data to disk or OS disk cache.
 	DiskSync disksync.Config `configKey:"diskSync"`
 	// DiskAllocation configures allocation of the disk space for file slices.
-	DiskAllocation allocate.Config `configKey:"diskAllocation"`
+	DiskAllocation diskalloc.Config `configKey:"diskAllocation"`
 }
 
 // ConfigPatch is same as the Config, but with optional/nullable fields.
@@ -24,7 +24,7 @@ type ConfigPatch struct {
 	// DiskSync configures the synchronization of the in-memory copy of written data to disk or OS disk cache.
 	DiskSync *disksync.Config `json:"diskSync,omitempty"`
 	// DiskAllocation configures allocation of the disk space for file slices.
-	DiskAllocation *allocate.Config `json:"diskAllocation,omitempty"`
+	DiskAllocation *diskalloc.Config `json:"diskAllocation,omitempty"`
 }
 
 // With copies values from the ConfigPatch, if any.
@@ -46,6 +46,6 @@ func NewConfig() Config {
 	return Config{
 		Compression:    compression.NewConfig(),
 		DiskSync:       disksync.NewConfig(),
-		DiskAllocation: allocate.NewConfig(),
+		DiskAllocation: diskalloc.NewConfig(),
 	}
 }
