+++ internal/pkg/service/buffer/sink/tablesink/storage/level/local/file.go
@@ -2,7 +2,7 @@ package local
 
 import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/compression"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/allocate"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/diskalloc"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/disksync"
 )
 
@@ -14,7 +14,7 @@ type File struct {
 	// DiskSync configures the synchronization of the in-memory copy of written data to disk or OS disk cache.
 	DiskSync disksync.Config `json:"diskSync"`
 	// DiskAllocation configures pre-allocation of the disk space for file slices.
-	DiskAllocation allocate.Config `json:"diskAllocation"`
+	DiskAllocation diskalloc.Config `json:"diskAllocation"`
 }
 
 func NewFile(cfg Config, fileDir string) File {
