+++ internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/diskalloc/allocator_other.go
@@ -1,6 +1,6 @@
 //go:build !linux
 
-package allocate
+package diskalloc
 
 import (
 	"os"
