+++ internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/diskalloc/config.go
@@ -1,4 +1,4 @@
-package allocate
+package diskalloc
 
 import "github.com/c2h5oh/datasize"
