+++ internal/pkg/service/buffer/sink/tablesink/storage/level/local/writer/diskalloc/config_test.go
@@ -1,4 +1,4 @@
-package allocate
+package diskalloc
 
 import (
 	"testing"
